// src/typings/ci.ts
export interface Environments {
  [propName: string]: string;
}

export interface StepsInterface {
  shells: string[];
  tab: string;
  script: string | null;

  addScriptCode: (s: string) => boolean;
  addShell: (s: string) => boolean;
  toString: () => string;
}

export interface StageInterface {
  tab: string;
  name: string;
  steps: StepsInterface | null;
  environments: Environments;

  // eslint-disable-next-line
  addEnvironment: (name: string, val: string) => void;
  addSteps: () => StepsInterface;
  toString: () => string;
}

export interface StagesInterface {
  tab: string;
  stages: StageInterface[];

  // eslint-disable-next-line
  addStage: (n: string) => StageInterface;
  toString: () => string;
}